Fix merge duplication
